+++ src/libs/zbxdbhigh/graph_linking.c
@@ -975,7 +975,7 @@ static void	update_graphs_items_updates(char **sql, size_t *sql_alloc, size_t *s
 							host_items_entry->drawtype_new);
 					d2 = ",";
 
-					zbx_audit_graph_update_json_update_gitem_update_drawtype(graphid, graph_flags,
+					zbx_audit_graph_update_json_update_gitem_update_drawtype(graphid, (int)graph_flags,
 							host_items_entry->gitemid, host_items_entry->drawtype_orig,
 							host_items_entry->drawtype_new);
 				}
@@ -986,7 +986,7 @@ static void	update_graphs_items_updates(char **sql, size_t *sql_alloc, size_t *s
 							host_items_entry->sortorder_new);
 					d2 = ",";
 
-					zbx_audit_graph_update_json_update_gitem_update_sortorder(graphid, graph_flags,
+					zbx_audit_graph_update_json_update_gitem_update_sortorder(graphid, (int)graph_flags,
 							host_items_entry->gitemid, host_items_entry->sortorder_orig,
 							host_items_entry->sortorder_new);
 				}
@@ -1000,7 +1000,7 @@ static void	update_graphs_items_updates(char **sql, size_t *sql_alloc, size_t *s
 					zbx_free(color_esc);
 					d2 = ",";
 
-					zbx_audit_graph_update_json_update_gitem_update_color(graphid, graph_flags,
+					zbx_audit_graph_update_json_update_gitem_update_color(graphid, (int)graph_flags,
 							host_items_entry->gitemid, host_items_entry->color_orig,
 							host_items_entry->color_new);
 				}
@@ -1011,7 +1011,7 @@ static void	update_graphs_items_updates(char **sql, size_t *sql_alloc, size_t *s
 							host_items_entry->yaxisside_new);
 					d2 = ",";
 
-					zbx_audit_graph_update_json_update_gitem_update_yaxisside(graphid, graph_flags,
+					zbx_audit_graph_update_json_update_gitem_update_yaxisside(graphid, (int)graph_flags,
 							host_items_entry->gitemid, host_items_entry->yaxisside_orig,
 							host_items_entry->yaxisside_new);
 				}
@@ -1022,7 +1022,7 @@ static void	update_graphs_items_updates(char **sql, size_t *sql_alloc, size_t *s
 							host_items_entry->calc_fnc_new);
 					d2 = ",";
 
-					zbx_audit_graph_update_json_update_gitem_update_calc_fnc(graphid, graph_flags,
+					zbx_audit_graph_update_json_update_gitem_update_calc_fnc(graphid, (int)graph_flags,
 							host_items_entry->gitemid, host_items_entry->calc_fnc_orig,
 							host_items_entry->calc_fnc_new);
 				}
@@ -1032,7 +1032,7 @@ static void	update_graphs_items_updates(char **sql, size_t *sql_alloc, size_t *s
 					zbx_snprintf_alloc(sql, sql_alloc, sql_offset, "%stype=%d", d2,
 							host_items_entry->type_new);
 
-					zbx_audit_graph_update_json_update_gitem_update_type(graphid, graph_flags,
+					zbx_audit_graph_update_json_update_gitem_update_type(graphid, (int)graph_flags,
 							host_items_entry->gitemid, host_items_entry->type_orig,
 							host_items_entry->type_new);
 				}
